Move platformInit and plugin load to server
