+++ cmd/containerd/main_linux.go
@@ -9,7 +9,6 @@ import (
 	"github.com/containerd/containerd/log"
 	"github.com/containerd/containerd/reaper"
 	"github.com/containerd/containerd/server"
-	"github.com/containerd/containerd/sys"
 )
 
 const defaultConfigPath = "/etc/containerd/config.toml"
@@ -19,22 +18,6 @@ var handledSignals = []os.Signal{
 	unix.SIGINT,
 	unix.SIGUSR1,
 	unix.SIGCHLD,
-}
-
-func platformInit(ctx context.Context, config *server.Config) error {
-	if config.Subreaper {
-		log.G(ctx).Info("setting subreaper...")
-		if err := sys.SetSubreaper(1); err != nil {
-			return err
-		}
-	}
-	if config.OOMScore != 0 {
-		log.G(ctx).Infof("changing OOM score to %d", config.OOMScore)
-		if err := sys.SetOOMScore(os.Getpid(), config.OOMScore); err != nil {
-			return err
-		}
-	}
-	return nil
 }
 
 func handleSignals(ctx context.Context, signals chan os.Signal, server *server.Server) error {
