+++ cmd/containerd/main_windows.go
@@ -18,10 +18,6 @@ var (
 	}
 )
 
-func platformInit(context context.Context, config *server.Config) error {
-	return nil
-}
-
 func handleSignals(ctx context.Context, signals chan os.Signal, server *server.Server) error {
 	for s := range signals {
 		log.G(ctx).WithField("signal", s).Debug("received signal")
