+++ server/server_unsupported.go
@@ -1,0 +1,9 @@
+// +build !linux
+
+package server
+
+import "context"
+
+func apply(_ context.Context, _ *Config) error {
+	return nil
+}
